Add Access methods on TwigRenderer
I searched any issue about this "problem" but I not found. 
I'm extending a Renderer, but my class CloudRenderer is very similar to TwigRenderer, so I extends this class. But, when I would rewrite the render method, I dont had access to defaultOptions, for example. I think that their methods will be needs in the future.