Refactored the RouterVoter to make it more flexible
The way to pass routes in the item extras has changed.

before:

    'routes' => array('foo', 'bar'),
    'routeParameters' => array('foo' => array('id' => 4))

after:

    'routes' => array(
        array('route' => 'foo', 'parameters' => array('id' => 4)),
	'bar',
    )

This new syntax allows using the same route with multiple parameter sets
for the same item, or matching the route for any parameter while using
some parameters to generate the url in the RouterAwareFactory by adding it
explicitly without parameters.

The old syntax is deprecated, and using it will trigger a
E_USER_DEPRECATED error.

Parameters are now checked against the _route_params attribute (introduced
in Symfony 2.1) instead of the attributes themselves, to avoid issues with
type casting when using param converters to replace the attributes by
objects. Closes #92 and replaces #93.